requirements.txt
numpy 
pandas
facenet-pytorch
faiss-cpu
opencv-python
torch
matplotlib
yaml